Create AvatarAndMenu component
Move the avatar and menu into a separate component to make PageHeader more manageable and not do more than it should.
